Limit Showcase mount to just development
Since Showcase is now a hard BulletTrain dependency, and since gemspec dependencies can't specify an environment,
we'd up mounting the Showcase routes in production. We shouldn't do that.

Due to a version of Showcase's tests using integration tests, we'd also mount the routes in the test environment,
which was the reason for using the defined? check.

We can now just use development?.