super_scaffolding depends on a newer version of Bullet Train
We're depending on a version that needs scaffolding_things_disabled? and silence_logs?
defined, so 1.0.x is too low and we'll blow up.

Depending on >= 1.2 shouldn't be a problem since these gems are all versioned the same.

So if you install super_scaffolding 1.2.15 you should already be on bullet_train 1.2.x,
and if not your app will throw exceptions so you'll have to update.