Refactor `with_attribute_settings`
Seems like we can just use one hash to store the settings in.
And then use keyword argument defaults to preserve the old behavior of keeping the previous setting in case
no override is passed.

This also uses `rescue` so we revert to the previous values in case an exception is thrown.