We should run the super_scaffolding_partial_test as well
Previously we could introduce changes in `core` that would break tests
in the starter repo, but that wouldn't show up as failures in `core`.

This was due to the fact that we were running more tests in the starter
repo than we were runnig here.

This PR makes it so that we run the extra tests so that we can spot test
failures here.