Generate clickable anchor links in `/docs` (#468)
Fixes #191

This relies on the `1.x` branch of the `commonmarker` gem, which is
currently in pre-relase stages.

We're waiting for `1.x` to get to a general release before we update our
dependencies and force the change on downstream users. The maintainer of
that gem has estimated that the general release should happen sometime
this fall, maybe around October. Once that release is available and once
we've updated our depencency list this change should Just Work.

In the meantime downstream applications can opt-in to using a
pre-release candidate of `commonmarker` by adding this to the `Gemfile`:

```ruby
gem 'commonmarker', '~> 1.0.0.pre10'
```