Make the showcase theme selector sticky
Adding `fixed` to the theme selector make it stick to the bottom of the
browser window instead of being stuck all the way at the bottom of the
scroll.

But if we _just_ add the `fixed` class, then the theme selector will
overlap the content at the very bottom of the page and we have no way to
get to it.

We can fix that by using a little bit of JS to calculate the height of
the theme selector and then set that amount of padding on the bottom of
the `main` container.

Doing that makes it so that there is a scrollbar if the theme selector
will overlap any content at the bottom of the `main` container.

And makes it so that you can scroll all the way down and see all of the
content.