Removes v2 api templates and partials
This is confusing.

The gem should only provide v1 versions and the templates and partials in the api/v2 dir are very out of date from the 'real' v1 versions.

If you are updating your application to a v2 or above, you would expect to provide the templates yourself, and not have out-of-date versions of few models provided.