Bug fix for super scaffolding an address_field (#600)
The PR for testing super scaffolding of Addresses (https://github.com/bullet-train-co/bullet_train/pull/986)
landed right before the PR for automatically generating a migration
during super scaffolding (https://github.com/bullet-train-co/bullet_train-core/pull/547).
As a result a bug / test-failure crept in.

The `Address` model `belongs_to :addressable`, which means that
the relation is stored on the `Address` side, and not on the owning
model. So that means that when we super scaffold an address field it
turns out that we don't need to represent that in the migration for
the model that will own the address.