[CI] Be more explicit about some dependencies (#667)
In the course of [trying to be able to run all of the tests for
individual gems in CI](https://github.com/bullet-train-co/bullet_train-core/pull/569) I've
discovered that we don't specify all of our dependencies fully.

In the starter repo the problems don't appear because all of the things
that we need end up being dependencies of something or other, but when
we try to run tests for individual gems we'll get errors about
unsatisfied dependencies.

There are only a few gems where we needed to declare a missing
dependency in the `.gemspec`. And those are the only thing in this PR
that could have the chance to impact downsteam apps.

All of the changes to `Gemfile` and `Gemfile.lock` in each gem won't
impact gem consumers at all. They just come into play in local
development and CI. The changes in each `Gemfile` ensure that when
developing or testing these gems that they'll use their peer gems from
inside the `core` repo insted of using whatever version happens to be
latest.