Fix: Handle internationalized git in superscaffold
Git commands can be internationalized and superscaffolding untracked file detection will fail because the labels don't match (eg, on my machine, git displays "Fichiers non suivis" instead of "Untracked files"), which causes the script to stop right after model generation.

I first modified `git_status`and `has_untracked_files`to use the --porcelain option of git status which, according to the docs, is intended to be used by scripts and as such is not internationalized, but digging deeper, it seems these are not needed, as `get_untracked_file` already manages the "no untracked files" scenario by returning an empty array.

As a workaround until this or a better solution is merged, one can re-run the last command with `--skip-migration-generation` and the script will successfully run this time