Don't throw an error for `_id` fields that don't incldue a `class_name` option
Fixes: https://github.com/bullet-train-co/bullet_train/issues/1456
Fixes: https://github.com/bullet-train-co/bullet_train/issues/1507

This makes it so that if someone super scaffolds a field that ends with
`_id` but does not explicitly include a `class_name` option then we'll
try to guess the `class_name` based on the name of the field.

For instance if you do this:

```
rails g super_scaffold Task Team event_id:super_select
```

We'll guess that `event_id` is a reference to the `Event` class. If we
can find an `Event` class in your app then everything proceeds as if you
had done this:

```
rails g super_scaffold Task Team event_id:super_select{class_name:Event}
```

If we can't find an `Event` model then you'll get a message saying so
and telling you about the `{vanilla}` option that you can use to disable
the `_id` related magic.