Raise an error if a namespaced model would clobber an existing class
Fixes: https://github.com/bullet-train-co/bullet_train/issues/1457

Previously superscaffolding commands would hang indefinitely if you
tried to geneate a model that has a namespace that exactly matches an
existing class.

Now we'll detect that problem and will show you a message letting you
know that you need to deconflict the namespace.

```
$ rails generate super_scaffold:join_model InputItem::PromptExecution prompt_execution_id{class_name=PromptExecution} input_item_id{class_name=InputItem}
It looks like the namespace you gave for this model conflicts with an existing class: InputItem
You should use a namespace that doesn't clobber an existing class.

We reccomend using the pluralized version of the existing class.

For instance instead of InputItem use InputItems
```