Prevent over-eager string replacement in factory during super scaffolding
Fixes: https://github.com/bullet-train-co/bullet_train/issues/1506

This change ensure that attributes named with a `_team` suffix don't get
clobbered in the factory by over eager string replacement.

For example if you do this:

```
bin/rails generate model Category team:references name:string owning_team:references
```

The first thing that happens is rails generates a model and a factory
and the factory looks like this:

```ruby
FactoryBot.define do
  factory :category do
    team { nil }
    name { "MyString" }
    owning_team { nil }
  end
end
```

Previously we'd end up trying to replace every instance of `team { nil }` with `association: team`.

That would end up breaking the line for `owning_team` by making it read
`owning_association :team`.

By including spaces in the pattern to be replaced (and the replacement)
we ensure that we only mess with the primary `team` relationship, and
not any other attributes.