Fix indentation on controller test generated by super scaffolding
In https://github.com/bullet-train-co/bullet_train/pull/1207 we changed
how the template controller test is structured. But we didn't updated
the super scaffolder to accomodate the change in structure.

Since we no longer wrap the entire test in `unless scaffolding_things_disabled?`
now we don't need to remove the indentation that was previously there due to that conditional.