Remove a redundant `.trix-content` div
Previously we were manually adding a `div.trix-content` element but we
only needed to do that because we didn't have a file in place that is
created when you run `rails action_text:install`. I've added that file
in the starter repo, where devs would normally expect to find it, so we
don't need to manually add the div anymore.

https://github.com/bullet-train-co/bullet_train/pull/1583/files#diff-77f493605c04a022c76dff16f1eca94e25e08854d33261e06ea0d7144302bf74