Make `two_factor_authentication_enabled?` aware of keys set via ENV vars
Fixes https://github.com/bullet-train-co/bullet_train-core/issues/895

Previously we only checked to see if a `primary_key` was set using Rails
encrypted credentials files. If it was set via ENV var then ActiveRecord
encryption would be enabled, but we'd incorrectly determine that it
wasn't enabled. This makes it so that two factor stuff shows up when
you've set the keys in the ENV.