Don't show the "invite new team member" button to people who aren't allowed to do that
By default BulletTrain ships with `roles.yml` configured such that any
user can invite other people to the team with equal or lesser
permissions than the inviter has.

If you change the `default` role so that they only have `read`
permissions for the `Invitation` model we still show the "Invite a New
Team Member" button to people with that role. If you click that button
(as a person with the `default` role) then we redirect you back to the
dashboard with an error flash message saying that you're not allowed to
access that page.

This PR makes it so that we don't show you the button at all if you're
not actually allowed to invite people.