Improved Heroku Detection
Our previous method of detecting that we're in Heroku required the app
to run the following command:

```
heroku labs:enable runtime-dyno-metadata
```

We didn't have that documented anywhere, so new apps often had trouble
around Redis. (It's super dumb that Heroku is **still** using janky
self-signed certs for their Redis offering.)

This PR iproves our detection method in a number of ways.

First we'll check for the existence of the `BT_IS_IN_HEROKU` ENV var.
This can be used to explicitly tell BT whether or not it's in Heroku.
We'll set this value to `true` in `app.json` of the Starter Repo. So new
apps that are deployed via `app.json` will have this set. This is the
safest and most stable method.

Next we'll do the check that we used to do for `HEROKU_APP_NAME`. We're
keeping it for backwards compatibility and because it's probably more
stable than the following fallbacks.

Finally we'll check for a couple of artifacts that Heroku currently
leaves laying around. If either the `DYNO` ENV var is present, or if
`/app/.heroku` exists, and is a directory, then we'll assume that we're
in Heroku. These artifacts may not be stable, so who knows how long
these fallbacks will work. Hopefully for a while. 🤞