Remove `extended_email_reply_parser` dependency (#1002)
* Remove `extended_email_reply_parser` dependency

For some reason `extended_email_reply_parser` was added as a dependency to the main `bullet_train` gem even though it's not used anywhere in `core` or in the starter repo.

It _is_ used by `bullet_train-conversations` and is declared as a dependency there, so we don't need it here.

`extended_email_reply_parser` relies on `charlock_holmes` which is notoriously finicky to install due to native extensions. Removing this dependency should eliminate one of the big problems that people often run into when first setting up Bullet Train.

Fixes https://github.com/bullet-train-co/bullet_train-core/issues/999

* Oops missed one

* Need to define another teplate method

* linter