Add more template methods (#1004)
* Add more template methods

Some of the billing gems don't have an `index` or `show` method, but
they use `account_load_and_authorize_resource`, which assumes that all
CRUD methods are always present on controllers that ues it.

This adds those template methods so that we don't throw an error when
accessing those controller in dev and test environments.

Related to: https://github.com/bullet-train-co/bullet_train-core/pull/930

* define a method to override the template

* fix a failing test