Allow user menu to be close after a click inside it
Previously if you clicked somewhere inside the user menu, without
clicking on an actual link it would effectively lock the menu into an
open state with no way to close it.

This makes it so that a click outside the menu will close it.

Fixes https://github.com/bullet-train-co/bullet_train-core/issues/987