Allow applications to declare their own root routes (#1013)
Previously if you tried to declare a public root route (without a unique
`as:` reference) you'd get an `ArgumentError` saying:

```
Invalid route name, already in use: 'root'
```

Similary if you tried to declare an account scoped dashboard route (with
an `as: 'dashboard'` reference) you'd get an `ArgumentError` saying:

```
Invalid route name, already in use: 'account_dashboard'
```

Both of these errors were happening when we tried to draw default routes
after the application had already declared them.

Now we resuce the `ArgumentError` and check the message to see if it's a
complaint about a duplicate route. If so, we do nothing and let the app
use the route defined in the primary `config/routes.rb` file.

Fixes https://github.com/bullet-train-co/bullet_train/issues/1656