Remove the `ruby-openai` gem
We included it as a dependency, and `require`d it, but we don't
actually use it anywhere other than in the `bullet_train-action_models`
gem. We should list it as a dependency there so that we don't force
literally every BT app to have it.

Fixes https://github.com/bullet-train-co/bullet_train-core/issues/1036