Remove the `image_processing` gem
We used to include `image_processing` as a hard dependency in the
`bullet_train` gem, but we don't use it directly (or even `require` it).

We recommend that people use Cloudinary for handling images, which means
that they don't need `image_processing` which is used for handling
images with `ActiveStorage`.

If you want to use `image_processing` then you should uncomment the line
for it in the `Gemfile` of your Bullet Train application.

Fixes: https://github.com/bullet-train-co/bullet_train-core/issues/1040