Remove `xxhash` as a hard dependency
We used to list `xxhash` as a hard dependency, but it was
only used in the generation of random colors for user & team icons.

If you have an existing app **and** you need to preserve the random
colors that used to be generated you can uncomment this gem and things
will continue to work the way they always have.

If you don't care about preserving the same random colors then you don't
need to include it.

Fixes https://github.com/bullet-train-co/bullet_train-core/issues/1049