Remove the `valid_email` gem
For some reason we listed this gem as a dependency, but we never
actually use it anywhere.

Since it was completely unused by default we're not adding it to the
"recommended but options gems" section of the `Gemfile` in the starter
repo.

If you want to use it you should follow the [instruction in the docs for
the gem](https://github.com/hallelujah/valid_email).

Fixes https://github.com/bullet-train-co/bullet_train-core/issues/1047