Remove `unicode-emoji` as a hard dependency (#1057)
* Remove `unicode-emoji` as a hard dependency

We used to declare `unicode-emoji` as a dependency, but it's only used
in one method that we don't seem to actually call anywhere.

That method has been updated to check if `unicode-emoji` is available.
If so it will proceed as normal, if not it will raise an error saying
that you should add `unicode-emoji` to your `Gemfile`.

Fixes https://github.com/bullet-train-co/bullet_train-core/issues/1048

* Add unicode-emoji for testing the gem, but not as a dependency