Hoist `rack-cors` into the `Gemfile` of the starter repo (#1061)
Previously we listed `rack-cors` as a hard dependency in the `core` gems, which meant that application developers were forced to include the gem even if they're not using the api. This makes it possible to remove `rack-cors` from the application completely.

Joint PR: https://github.com/bullet-train-co/bullet_train/pull/1954

Fixes: https://github.com/bullet-train-co/bullet_train-core/issues/1050