Use `Gem::Specification` instead of relying on shelling out to `bundle show` (#1236)
Apparently some systems/configurations can end up in a state where doing something like `bundle show bullet_train-themes-light` will result in bundler showing the location of the gem twice, which then messes up code that expects the output to be a single line.

So instead we'll use `Gem::Specification` direclty to find the info we need.